Create frontend/public/pages and move index.html, publicFeed.html, & register.html into it. Update all links and routes
Add link to publcFeed.html on index.html

Create frontend/public/modules

modularized createPostCard function -moved to modules

modularized handlePostSubmit function -moved to modules

modularized getFeed function -moved to modules
