+++ social-app/frontend/public/js/publicFeed.js
@@ -1,78 +1,5 @@
+import { loadFeed } from "/modules/loadFeed.js";
 //fetch on page load
 document.addEventListener('DOMContentLoaded', () => {
     loadFeed();
 });
-
-//GET posts from /api/posts/feed
-async function loadFeed() {
-
-    try {
-        const response = await fetch('/api/posts/feed', {
-            method: 'GET',
-        });
-        const posts = await response.json();
-
-        if (!posts.length) {
-
-            document.getElementById('emptyFeedMsg').hidden = false;
-            return;
-        }
-        //Loop through posts
-        let feedContainer = document.getElementById('feedContainer');
-        for (const post of posts) {
-            //Create the card 
-            let card = createPostCard(post);
-            //append postCard from createPostCard
-            feedContainer.append(card);
-        };
-
-    } catch (error) {
-        document.getElementById('feedFailMsg').hidden = false;
-    }
-}
-
-
-///helper loop function///
-
-//helper function loops through array, renders each post into DOM
-function createPostCard(post) {
-    //create div
-    let divCard = document.createElement('div');
-    divCard.className = 'postCard';
-
-    //create username h3
-    let username = document.createElement('h3');
-    username.className = 'postcardUsername';
-    username.innerText = post.author.username;
-    divCard.append(username);
-
-    //create add post content
-    let content = document.createElement('p');
-    content.className = 'cardContent';
-    content.innerText = post.content;
-    divCard.append(content);
-
-    //check if post.mediaUrl exists and is not empty
-    if (post.mediaUrl) {
-        //if yes create img element, set src to mediaUrl, give class postMedia
-        let postMedia = document.createElement('img');
-        postMedia.className = 'postMedia';
-        postMedia.src = post.mediaUrl;
-        divCard.append(postMedia)
-    }
-
-    //create formatted timestamp
-    let cardTime = document.createElement('sub')
-    cardTime.className = 'cardTimestamp';
-    //get createdAT
-    let createdAt = post.createdAt;
-    //convert from string to Date object
-    let date = new Date(createdAt);
-    //format date into a readable string
-    let cardPostDate = date.toDateString();
-    cardTime.innerText = cardPostDate;
-    divCard.append(cardTime);
-    //return back inside loadfeed() append this result to feedContainer
-    return divCard;
-
-}
