+++ social-app/frontend/public/pages/index.html
@@ -4,8 +4,8 @@
 <head>
     <meta charset="UTF-8" />
     <title>Social Login</title>
-    <link rel="stylesheet" href="css/index.css">
-    <script type="module" src="js/login.js"></script>
+    <link rel="stylesheet" href="/css/index.css">
+    <script type="module" src="/js/login.js"></script>
 
 <body>
 
@@ -30,7 +30,8 @@
             </div>
             <div>
                 <button class="formButton" type="submit"> Login </button>
-                <p id="reg_link"> or Create an account <a href="/register.html">Here</a>. </p>
+                <p><a href="publicFeed.html" id="pub_link">Browse public</a> or Create an account <a
+                        href="register.html" id="reg_link">Here</a>. </p>
             </div>
         </form>
