+++ social-app/frontend/public/pages/publicFeed.html
@@ -4,9 +4,9 @@
 <head>
     <meta charset="UTF-8" />
     <title>Public Feed</title>
-    <link rel="stylesheet" href="css/publicFeed.css">
-    <script src="js/publicFeed.js"></script>
-    <script type="module" src="js/login.js"></script>
+    <link rel="stylesheet" href="/css/publicFeed.css">
+    <script type="module" src="/js/publicFeed.js"></script>
+    <script type="module" src="/js/login.js"></script>
 
 </head>
 
@@ -35,7 +35,7 @@
                 </div>
                 <div>
                     <button class="formButton" type="submit"> Login </button>
-                    <p id="reg_link"> or Create an account <a href="/register.html">Here</a>. </p>
+                    <p id="reg_link"> or Create an account <a href="register.html">Here</a>. </p>
                 </div>
             </form>
