+++ social-app/frontend/public/pages/register.html
@@ -31,7 +31,7 @@
             <div>
                 <button class="formButton" type="submit"> Register </button>
             </div>
-            <p id="reg_link"> Return to <a href="/index.html">Login</a>. </p>
+            <p id="reg_link"> Return to <a href="index.html">Login</a>. </p>
         </form>
 
         <p id="registerError" class="errorMessage" hidden> Invalid Credentials</p>
